+++ CHANGELOG.md
@@ -9,8 +9,9 @@ All Notable changes to `Csv` will be documented in this file
 - More return types and type parameters as supported in PHP7.2+
 - `League\Csv\Statement::create` named constructor to ease constraint builder instantiation
 - `League\Csv\Statement` can now also process `League\Csv\ResultSet` instances.
-- `League\Csv\ResultSet::getRecords` has an optional `$header` second argument to make the method works like `League\Csv\Reader::getRecords` 
 - `League\Csv\TabularDataReader` interface to represent how to read tabular data
+- `League\Csv\ResultSet::getRecords` has an optional `$header` second argument to make the method works like `League\Csv\Reader::getRecords` 
+- `League\Csv\ResultSet::createFromReader` create a new instance from `League\Csv\Reader` 
 
 ### Deprecated